refactoring modules for .mjs extension and include babel to generate commonjs build
