+++ examples/helloworld-cjs/package.json
@@ -4,13 +4,12 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node index.js"
   },
   "keywords": [],
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "createmeal": "1.0.1",
-    "express": "^4.17.1"
+    "express": "^4.17.3"
   }
 }
