+++ examples/helloworld-npm/package.json
@@ -4,14 +4,13 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node index.mjs"
   },
   "keywords": [],
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "createmeal": "^1.0.0",
-    "express": "^4.17.1"
-  },
-  "type": "module"
+    "createmeal": "^2.1.0",
+    "express": "^4.17.3"
+  }
 }
