+++ examples/helloworld-npm/public/index.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html><head title="Hello World NPM"></head><body><div class="container"><h3>Hello World!</h3><p>This is a pre-rendered file.</p></div></body></html>
+<!DOCTYPE html><html><head><title>Hello World NPM</title></head><body><div class="container"><h3>Hello World!</h3><p>This is a pre-rendered file.</p></div></body></html>
