+++ examples/helloworld-ts/package.json
@@ -6,16 +6,16 @@
   "scripts": {
     "build": "tsc --build",
     "debug": "tsc --sourcemap",
-    "start": "ts-node index.ts"
+    "start": "ts-node ./index.ts"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@types/node": "^17.0.10",
-    "createmeal": "^1.1.0",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.5.4"
+    "@types/node": "^17.0.21",
+    "createmeal": "^2.1.0",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.2"
   },
   "type": "module"
 }
