+++ examples/helloworld-ts/tsconfig.json
@@ -1,10 +1,14 @@
 {
   "compilerOptions": {
     "target": "es6",
-    "module": "es2020",
+    "module": "ESNext",
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
-    "sourceMap": true
+    "sourceMap": true,
+    "moduleResolution": "node"
+  },
+  "ts-node": {
+      "esm": true
   }
 }
