+++ package.json
@@ -2,17 +2,17 @@
   "name": "createmeal",
   "version": "2.0.0",
   "description": "A library to render HTML",
-  "main": "dist/createmeal.js",
+  "main": "dist/createmeal.mjs",
   "scripts": {
     "eslint": "eslint ./src/**/*",
     "build": "webpack",
     "build-tsc": "tsc --build",
-    "docs": "node ./documentation/readme-generator.js",
+    "docs": "node ./documentation/readme-generator.mjs",
     "test": "c8 mocha ./test/**/*.js",
     "release": "node release.js",
-    "release_patch": "node release.js patch",
-    "release_minor": "node release.js minor",
-    "release_major": "node release.js major",
+    "release_patch": "node release.mjs patch",
+    "release_minor": "node release.mjs minor",
+    "release_major": "node release.mjs major",
     "git_tag": "git tag -a $(echo $(npm pkg get version) | sed -e 's/\"//g') -m 'release new version'"
   },
   "author": "Julio Cezar da Silva (https://github.com/jcezardasilva)",
@@ -36,6 +36,10 @@
   ],
   "license": "MIT",
   "devDependencies": {
+    "@babel/core": "^7.17.5",
+    "@babel/plugin-transform-modules-commonjs": "^7.16.8",
+    "@babel/preset-env": "^7.16.11",
+    "babel-loader": "^8.2.3",
     "c8": "^7.11.0",
     "chai": "^4.3.6",
     "eslint": "^8.9.0",
@@ -50,7 +54,6 @@
   "directories": {
     "test": "test"
   },
-  "type": "module",
   "c8": {
     "src": [
       "/src"
