+++ src/composite/SelfClosingTag.mjs
@@ -1,4 +1,4 @@
-import Node from "./Node.js";
+import Node from "./Node.mjs";
 export default class SelfClosingTag extends Node {
     constructor(name){
         super(name);
