+++ src/selfClosingTags.mjs
