+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   // Change this to match your project
-  "include": ["dist/*.js"],
+  "include": ["dist/createmeal-legacy-node.js","dist/createmeal-legacy.js","dist/createmeal.mjs"],
   "compilerOptions": {
     // Tells TypeScript to read JS files, as
     // normally they are ignored as source files
